# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Atiende Rentas Vacacionales — API
  version: 0.3.0
  description: |
    Contrato de API de `apps/api` (Lote 3: auth/roles/RLS/auditoría sobre el
    dominio de calendario de Lote 1). Fuente de verdad ejecutable: los
    esquemas `zod` de `apps/api/src/contrato/tipos.ts` (importables como
    `@atiende-rv/api/contrato`) — este archivo es la descripción
    human-readable del mismo contrato, publicada en el primer commit de
    este lote para que Lote 4 (frontend) pueda empezar contra tipos
    estables antes de que todos los endpoints tengan implementación
    completa.

    Reglas de dominio no negociables reflejadas en este contrato (D-006,
    D-011, BLUEPRINT): la API NUNCA expone cancelación ni modificación de
    una reserva cuyo origen sea un canal externo (Airbnb/Vrbo/Booking) —
    solo reservas DIRECTAS (creadas por esta misma API). Tampoco expone
    envío de mensajes a huéspedes (eso es Lote 6, con aprobación humana
    obligatoria). `POST /canales/:id/sync` solo ENCOLA una solicitud de
    sincronización manual en `outbox_evento`; la ejecución real la hace el
    worker de Lote 2.
servers:
  - url: http://localhost:8787
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /canales/cuentas:
    post:
      tags: [canales]
      summary: Registrar una cuenta de canal — admin_gestora/superadmin (credenciales cifradas AES-256-GCM)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearCuentaCanal" }
      responses:
        "201": { description: Cuenta creada }
        "403": { $ref: "#/components/responses/RolForbidden" }
...
  /auditoria:
    get:
      tags: [auditoria]
      summary: Consulta paginada del log de auditoría — admin_gestora/superadmin
      parameters:
        - name: pagina
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: tamano
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Página de entradas de auditoría
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RespuestaAuditoriaPaginada" }
        "403": { $ref: "#/components/responses/RolForbidden" }

  # ---------------------------------------------------------------------
  # Operación: limpieza/mantenimiento/inspección (Lote 5, BACKLOG E08).
  # Rol `limpieza` solo ve/opera sus propias tareas asignadas (RLS,
  # packages/db migración 0036); propietario tiene solo lectura de la
  # operación de sus propias unidades.
  # ---------------------------------------------------------------------
...
  /backoffice/flags:
    get:
      tags: [backoffice]
      summary: Catálogo de feature flags con valor global/por-tenant — superadmin
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Catálogo de flags
          content:
            application/json:
              schema:
                type: object
                properties:
                  flags:
                    type: array
                    items: { $ref: "#/components/schemas/Flag" }
  /backoffice/flags/{id}:
    patch:
      tags: [backoffice]
      summary: Cambiar el valor de un flag (global o por tenant), con motivo auditado — superadmin
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoEstablecerFlag" }
      responses:
        "200": { description: Flag actualizado }
        "404": { $ref: "#/components/responses/NoEncontrado" }
  /backoffice/flags/{id}/auditoria:
    get:
      tags: [backoffice]
      summary: Historial de cambios de un flag — superadmin
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Entradas de auditoría del flag }
        "404": { $ref: "#/components/responses/NoEncontrado" }
...
  /backoffice/cuentas-canal:
    get:
      tags: [backoffice]
      summary: Cuentas de canal del tenant — nunca expone credenciales, solo "configurada/no configurada"
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Cuentas de canal del tenant }
    post:
      tags: [backoffice]
      summary: Alta con tipo de conexión honesto — iCal / partner pendiente (con motivo) / simulador (solo dev)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CuerpoCrearCuentaCanalBackoffice" }
      responses:
        "201": { description: Cuenta creada }
        "422": { $ref: "#/components/responses/Validacion" }
...
  /backoffice/auditoria:
    get:
      tags: [backoffice]
      summary: Auditoría consultable con filtros (tabla/operación/actor/rango de fechas)
      parameters:
        - name: pagina
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: tamano
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: tabla
          in: query
          schema: { type: string }
        - name: operacion
          in: query
          schema: { type: string, enum: [INSERT, UPDATE, DELETE, ACCESO_ROMPER_CRISTAL] }
        - name: actorId
          in: query
          schema: { type: string, format: uuid }
        - name: desde
          in: query
          schema: { type: string, format: date-time }
        - name: hasta
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Página de entradas de auditoría filtradas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RespuestaAuditoriaPaginada" }
        "403": { $ref: "#/components/responses/RolForbidden" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Validacion:
      description: Error de validación de entrada (zod)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
...
    RolForbidden:
      description: "rol_forbidden: el rol/nivel de colaborador no tiene permiso para esta acción"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
    NoEncontrado:
      description: "recurso_no_encontrado (incluye recursos de otro tenant, ocultos por RLS — nunca confirma su existencia)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorHttp" }
  schemas:
    ErrorHttp:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [codigo, mensaje]
          properties:
            codigo:
              type: string
              enum:
                - credenciales_invalidas
                - token_invalido
                - token_expirado
                - tenant_forbidden
                - rol_forbidden
                - rango_invalido
                - unidad_no_disponible
                - conflicto_pendiente
                - reserva_no_directa
                - recurso_no_encontrado
                - validacion
                - rate_limited
                - error_interno
            mensaje: { type: string }
            detalles: {}
...
    CuerpoCrearCuentaCanal:
      type: object
      required: [canalCodigo, nombre]
      properties:
        tenantId: { type: string, format: uuid }
        propiedadId: { type: string, format: uuid }
        canalCodigo: { type: string, enum: [airbnb, vrbo, booking, manual] }
        nombre: { type: string }
        credenciales:
          type: object
          additionalProperties: { type: string }
          description: Se cifra con AES-256-GCM antes de persistir (H-046) — nunca se guarda en claro.
        esSimulador: { type: boolean }
...
    EntradaAuditoria:
      type: object
      properties:
        id: { type: string, format: uuid }
        tabla: { type: string }
        filaId: { type: string, format: uuid }
        operacion: { type: string, enum: [INSERT, UPDATE, DELETE, ACCESO_ROMPER_CRISTAL] }
        actorId: { type: string, format: uuid, nullable: true }
        creadoEn: { type: string, format: date-time }
        valoresPrevios: { nullable: true }
        valoresNuevos: { nullable: true }
    RespuestaAuditoriaPaginada:
      type: object
      properties:
        pagina: { type: integer }
        tamano: { type: integer }
        total: { type: integer }
        entradas:
          type: array
          items: { $ref: "#/components/schemas/EntradaAuditoria" }

    # -----------------------------------------------------------------
    # Operación: limpieza/mantenimiento/inspección (Lote 5, BACKLOG E08)
    # -----------------------------------------------------------------
...
    CuerpoEstablecerFlag:
      type: object
      required: [valor, motivo]
      properties:
        valor: { type: boolean }
        tenantId: { type: string, format: uuid }
        motivo: { type: string, minLength: 1 }
    Flag:
      type: object
      properties:
        id: { type: string }
        descripcion: { type: string }
        categoriaRiesgo: { type: string }
        valorGlobal: { type: boolean }
        valorEfectivo: { type: boolean }
        overrideDeTenant: { type: boolean }
...
    CuerpoCrearCuentaCanalBackoffice:
      type: object
      required: [canalCodigo, nombre, tipoConexion]
      properties:
        tenantId: { type: string, format: uuid }
        propiedadId: { type: string, format: uuid }
        canalCodigo: { type: string, enum: [airbnb, vrbo, booking, manual] }
        nombre: { type: string }
        tipoConexion: { type: string, enum: [ical, partner_pendiente, simulador] }
        motivoPartnerPendiente: { type: string, description: "Obligatorio si tipoConexion=partner_pendiente" }
        urlImport: { type: string, format: uri }
        credenciales: { type: object, additionalProperties: { type: string } }
...